Revealing code problems

+++ training.py
@@ -1,4 +1,4 @@
-import json
+import copy
 import logging
 import os
 import pickle
@@ -15,7 +15,7 @@ class MyStringExporter(pgn.StringExporter):
 
     def __init__(self, comments):
         super().__init__(headers=True, variations=True, comments=True)
-        self.comm_stack = comments
+        self.comm_stack = copy.copy(comments)
 
     def visit_move(self, board, move):
         if self.variations or not self.variation_depth:
@@ -31,42 +31,40 @@ class MyStringExporter(pgn.StringExporter):
             self.force_movenumber = False
 
 
-def record_results(brd, rnd, avgs):
-    journal = pgn.Game.from_board(brd)
-    journal.headers.clear()
-    journal.headers["White"] = "Lisa"
-    journal.headers["Black"] = "Karen"
-    journal.headers["Round"] = rnd
-    journal.headers["Result"] = brd.result(claim_draw=True)
-    if brd.is_checkmate():
-        comm = "checkmate"
-    elif brd.can_claim_fifty_moves():
-        comm = "50 moves"
-    elif brd.can_claim_threefold_repetition():
-        comm = "threefold"
-    elif brd.is_insufficient_material():
-        comm = "material"
-    elif not any(brd.generate_legal_moves()):
-        comm = "stalemate"
-    else:
-        comm = "by other reason"
-    journal.headers["Site"] = comm
-
-    exporter = MyStringExporter(brd.comment_stack)
-    pgns = journal.accept(exporter)
-    # logging.info("\n%s", pgns)
-    logging.info("Game #%d:\t%s by %s,\t%d moves,\t%.3f AMS", rnd, journal.headers["Result"], comm,
-                 brd.fullmove_number, avgs)
-    with open("last.pgn", "w") as out:
-        out.write(pgns)
-
-
 class BoardOptim(Board):
 
     def __init__(self, fen=STARTING_FEN, *, chess960=False):
         super().__init__(fen, chess960=chess960)
         self._fens = []
         self.comment_stack = []
+
+    def write_pgn(self, fname, roundd):
+        journal = pgn.Game.from_board(self)
+        journal.headers.clear()
+        journal.headers["White"] = "Lisa"
+        journal.headers["Black"] = "Karen"
+        journal.headers["Round"] = roundd
+        journal.headers["Result"] = self.result(claim_draw=True)
+        journal.headers["Site"] = self.explain()
+        exporter = MyStringExporter(self.comment_stack)
+        pgns = journal.accept(exporter)
+        with open(fname, "w") as out:
+            out.write(pgns)
+
+    def explain(self):
+        if self.is_checkmate():
+            comm = "checkmate"
+        elif self.can_claim_fifty_moves():
+            comm = "50 moves"
+        elif self.can_claim_threefold_repetition():
+            comm = "threefold"
+        elif self.is_insufficient_material():
+            comm = "material"
+        elif not any(self.generate_legal_moves()):
+            comm = "stalemate"
+        else:
+            comm = "by other reason"
+        return comm
 
     def can_claim_threefold_repetition1(self):
         # repetition = super().can_claim_threefold_repetition()
@@ -92,13 +90,18 @@ def play_one_game(pwhite, pblack, rnd):
     pwhite.board = board
     pblack.board = board
 
-    while pwhite.makes_move() and pblack.makes_move():  # and board.fullmove_number < 150
-        # logging.debug("%s. %s %s", board.fullmove_number - 1, board.move_stack[-1], board.move_stack[-2])
-        pass
+    while True:  # and board.fullmove_number < 150
+        if not pwhite.makes_move():
+            break
+        if not pblack.makes_move():
+            break
+
+    board.write_pgn(os.path.join(os.path.dirname(__file__), "last.pgn"))
 
     all_moves = pwhite.moves_log + pblack.moves_log
     avg_score = sum([x.get_score() for x in all_moves]) / float(len(all_moves))
-    record_results(board, rnd, avg_score)
+    logging.info("Game #%d:\t%s by %s,\t%d moves,\t%.3f AMS", rnd, board.result(claim_draw=True), board.explain(),
+                 board.fullmove_number, avg_score)
 
 
 if __name__ == "__main__":
@@ -112,7 +115,9 @@ if __name__ == "__main__":
     dataset = set()
     if os.path.exists("moves.pkl"):
         with open("moves.pkl", 'rb') as fhd:
-            dataset.update(pickle.load(fhd))
+            loaded = pickle.load(fhd)
+            dataset.update(loaded)
+            nn.learn(dataset, 50)
 
     rnd = 0
     while True:
@@ -122,15 +127,20 @@ if __name__ == "__main__":
         wmoves = white.get_moves()
         bmoves = black.get_moves()
         game_data = wmoves + bmoves
+        l1 = len(dataset)
         dataset.update(game_data)
+        # logging.info("%s+%s=%s", l1, len(game_data), len(dataset))
+        assert len(dataset) < l1 + len(game_data)
 
         if not (rnd % 20):
-            os.rename("moves.pkl", "moves.bak.pkl")
+            if os.path.exists("moves.pkl"):
+                os.rename("moves.pkl", "moves.bak.pkl")
             try:
                 with open("moves.pkl", "wb") as fhd:
-                    pickle.dump(list(dataset), fhd)
+                    pickle.dump(dataset, fhd)
             except:
                 os.rename("moves.bak.pkl", "moves.pkl")
 
+            # break
             nn.learn(dataset, 10)
             nn.save("nn.hdf5")
